Fixed POSIX !searchparse bug (patch #251) and hopefully better compatibility with 2.46
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@6485 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/lang.cpp
@@ -848,7 +848,7 @@ static void CreatePlatformStrfmt(const TCHAR *templ, TCHAR *out) {
     if (*out == L'%' && *templ == L's')
     {
       out++, templ++;
-      unsigned int cch = my_strncpy(out, NPRIs, -1);
+      size_t cch = my_strncpy(out, NPRIs, -1);
       out += --cch; // --cch because for loop does out++
     }
   }
